Refactor JS part of parsing CqlValueWrapper
Moved the logic for creating result object to separate file
Mostly changes from 0af57c8b9beab7809c02386e3fe2b52444494a14
split into a separate PR for clarity reasons

Update the way empty result is handled
