+++ lib/types/result-set.js
@@ -4,6 +4,7 @@ const utils = require("../utils");
 const errors = require("../errors");
 const rust = require("../../index");
 const Row = require("./row");
+const { getRowsFromResultsWrapper } = require("./results-wrapper");
 
 const asyncIteratorSymbol = Symbol.asyncIterator || "@@asyncIterator";
 
@@ -130,46 +131,6 @@ const asyncIteratorSymbol = Symbol.asyncIterator || "@@asyncIterator";
 // }
 
 /**
- *
- * @param {rust.CqlValueWrapper} field
- * @returns {any}
- */
-function getCqlObject(field) {
-    let type = field.getType();
-    switch (type) {
-        case rust.CqlType.Ascii:
-            return field.getAscii();
-        case rust.CqlType.Blob:
-            return field.getBlob();
-        case rust.CqlType.Boolean:
-            return field.getBoolean();
-        case rust.CqlType.Counter:
-            return field.getCounter();
-        case rust.CqlType.Double:
-            return field.getDouble();
-        case rust.CqlType.Float:
-            return field.getFloat();
-        case rust.CqlType.Int:
-            return field.getInt();
-        case rust.CqlType.Set:
-            let res = [];
-            let fields = field.getSet();
-            for (let i = 0; i < fields.length; i++)
-                res.push(getCqlObject(fields[i]));
-            return res;
-        case rust.CqlType.Text:
-            return field.getText();
-        case rust.CqlType.SmallInt:
-            return field.getSmallInt();
-        case rust.CqlType.TinyInt:
-            return field.getTinyInt();
-
-        default:
-            return field.stringify();
-    }
-}
-
-/**
  * @class
  * @classdesc Represents the result of a query.
  * @param {rust.QueryResultWrapper} result
@@ -177,17 +138,7 @@ function getCqlObject(field) {
  */
 
 function ResultSet(result) {
-    let rust_rows = result.getRows();
-    let col_names = result.getColumnsNames();
-    this.rows = [];
-    for (let i = 0; i < rust_rows.length; i++) {
-        let cols = rust_rows[i].getColumns();
-        let cols_sum = {};
-        for (let j = 0; j < cols.length; j++) {
-            cols_sum[col_names[j]] = getCqlObject(cols[j]);
-        }
-        this.rows.push(cols_sum);
-    }
+    this.rows = getRowsFromResultsWrapper(result);
 }
 
 /**
